Traktor: Reduced size of some Flash classes. Added a "depth" argument to createExternalMovieClipInstance method allows movieclips to be created at a specified depth without using a container clip.
[git-p4: depot-paths = "//depot/traktor/": change = 20298]

+++ code/Animation/Animation/Animation.cpp
@@ -85,7 +85,7 @@ uint32_t Animation::addKeyPose(const KeyPose& pose)
 void Animation::removeKeyPose(uint32_t poseIndex)
 {
 	T_ASSERT (poseIndex >= 0 && poseIndex < uint32_t(m_poses.size()));
-	m_poses.erase(m_poses.begin() + poseIndex);
+	m_poses.erase(m_poses.begin() + size_t(poseIndex));
 }
 
 bool Animation::empty() const
