+++ code/Animation/Pose.cpp
@@ -71,12 +71,12 @@ const Pose::Joint* Pose::getJoint(uint32_t jointIndex) const
 
 Pose::Joint& Pose::getEditJoint(uint32_t jointIndex)
 {
-	uint32_t s = 0;
-	uint32_t e = uint32_t(m_joints.size());
+	size_t s = 0;
+	size_t e = m_joints.size();
 
 	while (s < e)
 	{
-		uint32_t m = s + (e - s) / 2;
+		size_t m = s + (e - s) / 2;
 		if (jointIndex == m_joints[m].index)
 			return m_joints[m];
 		else if (jointIndex < m_joints[m].index)
